✨ feat: CD 추가 API 연결
